requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
compressed-rtf==1.0.7
confection==0.1.5
cryptography==46.0.1
cymem==2.0.11
distro==1.9.0
docx2txt==0.9
ebcdic==1.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
extract-msg==0.28.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.6
jiter==0.11.0
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.3
olefile==0.47
openai==1.108.1
packaging==25.0
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==11.3.0
preshed==3.0.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.0
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==0.6.23
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.12.0
smart_open==7.3.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.5.1
textract==1.6.5
thinc==8.3.6
tqdm==4.67.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
xlrd==1.2.0
xlsxwriter==3.2.9
